split weather view and rename
