+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import 'package:flutter_training/views/weather_view.dart';
+
 void main() {
   runApp(const MainApp());
 }
@@ -11,88 +13,10 @@ class MainApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return const MaterialApp(
       home: Scaffold(
-        body: WeatherWidget(),
+        body: WeatherView(),
       ),
     );
   }
 }
 
-class WeatherWidget extends StatelessWidget {
-  const WeatherWidget({super.key});
 
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox.expand(
-      child: FractionallySizedBox(
-        widthFactor: 0.5,
-        child: Column(
-          children: [
-            const Spacer(),
-            Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                const AspectRatio(aspectRatio: 1, child: Placeholder()),
-                Padding(
-                  padding: const EdgeInsets.symmetric(vertical: 16),
-                  child: Row(
-                    children: [
-                      Expanded(
-                        child: Text(
-                          '** ℃',
-                          textAlign: TextAlign.center,
-                          style: Theme.of(context)
-                              .textTheme
-                              .labelLarge!
-                              .copyWith(color: Colors.blue),
-                        ),
-                      ),
-                      Expanded(
-                        child: Text(
-                          '** ℃',
-                          textAlign: TextAlign.center,
-                          style: Theme.of(context)
-                              .textTheme
-                              .labelLarge!
-                              .copyWith(color: Colors.red),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-            Expanded(
-              child: Column(
-                children: [
-                  const SizedBox(height: 80),
-                  Row(
-                    children: [
-                      Expanded(
-                        child: TextButton(
-                          onPressed: () {},
-                          child: const Text(
-                            'Close',
-                            textAlign: TextAlign.center,
-                          ),
-                        ),
-                      ),
-                      Expanded(
-                        child: TextButton(
-                          onPressed: () {},
-                          child: const Text(
-                            'Reload',
-                            textAlign: TextAlign.center,
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
